+++ design-system/stories/textFields/BoxTextField.stories.tsx
@@ -1,16 +1,16 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import DefaultTextField from '../../ui/textField/DefaultTextField';
+import BoxTextField from '../../ui/textFields/BoxTextField';
 
 const meta = {
-  title: 'TextFields/DefaultTextField',
-  component: DefaultTextField,
+  title: 'TextFields/BoxTextField',
+  component: BoxTextField,
   parameters: {
     layout: 'centered',
   },
   decorators: [
     Story => (
-      <div style={{ width: '350px' }}>
+      <div style={{ width: '380px' }}>
         <Story />
       </div>
     ),
@@ -22,24 +22,19 @@ const meta = {
       description: '텍스트 필드에 들어갈 문자열의 정보',
       defaultValue: '입력해주세요',
     },
-    type: {
-      control: { type: 'select', options: ['text', 'password'] },
-      description: '텍스트 필드의 타입',
-      defaultValue: 'text',
-    },
     onChange: {
       description: '입력 이벤트',
     },
   },
-} satisfies Meta<typeof DefaultTextField>;
+} satisfies Meta<typeof BoxTextField>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
 export const Default: Story = {
   args: {
-    label: 'label',
     onChange: e => console.log(e.target.value),
     placeholder: '입력해주세요',
+    className: 'h-12',
   },
 };
