+++ src/pages/join/InfoInputPage.tsx
@@ -3,7 +3,7 @@ import { useForm, SubmitHandler } from 'react-hook-form';
 import Header from '../../../design-system/ui/Header';
 import Button from '../../../design-system/ui/Button';
 import { validations } from '../../shared/utils/validation';
-import DefaultTextField from '../../../design-system/ui/textField/DefaultTextField';
+import DefaultTextField from '../../../design-system/ui/textFields/DefaultTextField';
 
 interface FormInputs {
   name: string;
